Updated PrintPreview Cancel to focus on main form
I update the Cancel button on the Print Preview to focus back on the
main form.  The problem was that if the cancel button was clicked the
main form didn't have focus so the hot keys didn't work.  You had to
click in the main text box to get the focus again.  This behavior caused
some automation software not to work well after a print preview was
done.