Added check for shift and support for numpad '+'
Fix for Issue #111.  Fix was in same area I had already been looking at
for the previous issue.  Checks for shift and looks for numpad + key.
Note that this will only work on US keyboards.  Since the rest of the
code is non-internationalized, this should probably be OK for now.
Also, file was auto-formatted by Studio.  If this is an issue I can
revert and redo the change and prevent it from formatting.