Additional fix for intellisense - Issues #111
When I was experimenting with determining the distinction between '@'
and '2', I altered the xaml to change Intellisense_KeyDown to use the
PreviewKeyDown instead of the PreviewKeyUp. This was to ensure that the
"ModifiedKeys" would always detect if shift was pressed.  Changing it
back, it now will mostly work: if you try to quickly let go of shift as
soon as 2 or = is pressed, it will not popup the intellisense.  For now,
it works well enough and would need to be addressed again anyhow if
international support is added.