Fix bugs in Postpone task functionality (Closes Issue #188)
Added the PostponeTask method to MainWindowViewModel.cs, which replaces
the prior implementation of the postpone task feature. This resolves two
bugs in the UI: (1) postponing a task in the task list would delete
whatever text was in the new/edit task text box; and (2) postponing a
task with a prepended date equal to the due date would update the task's
prepended date as well as the due date.

Closes Issue #188 (Postponing a task deletes whatever text is in the
new/update task text box).