Add Font Size Text Box to Font Selection Window (Addresses Issue #157)
This change addresses and potentially closes Issue #157 (Font size show
as float number).

I added a text field for font size into the font selection dialog. The
font size text box is bound to the slider in a two-way sync. With this
setup, the user can see and edit the exact font size in the text box,
and still use the slider to set fonts, and still have no restrictions on
the float value of the font size. This would address issue #157 because
the exact font size displayed and set in the font selection window would
now reflected in the options window. It also makes it easier to set
integer-value font sizes.

I also changed the slider's "big jump" value from 2 to 1 to make it
simpler to move the slider up and down by one point (by clicking on the
line). Lastly, I changed the slider minimum from 8 to 1 because the
minimum of 8 made the font size text box reject the first digit in font
sizes such as "12" and "25" because the first number was less than 8.